+++ languagetool-commandline/src/test/java/org/languagetool/commandline/MainTest.java
@@ -18,7 +18,7 @@
  */
 package org.languagetool.commandline;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.junit.After;
 import org.junit.Before;
 import org.junit.Ignore;