+++ languagetool-core/src/main/java/org/languagetool/JLanguageTool.java
@@ -18,7 +18,7 @@
  */
 package org.languagetool;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.jetbrains.annotations.Nullable;
 import org.languagetool.databroker.DefaultResourceDataBroker;
 import org.languagetool.databroker.ResourceDataBroker;