+++ languagetool-core/src/main/java/org/languagetool/languagemodel/LuceneSingleIndexLanguageModel.java
@@ -18,8 +18,9 @@
  */
 package org.languagetool.languagemodel;
 
-import org.apache.commons.lang.StringUtils;
-import org.apache.lucene.index.*;
+import org.apache.lucene.index.DirectoryReader;
+import org.apache.lucene.index.IndexReader;
+import org.apache.lucene.index.Term;
 import org.apache.lucene.search.*;
 import org.apache.lucene.store.FSDirectory;
 import org.languagetool.Experimental;
@@ -127,7 +128,7 @@ public class LuceneSingleIndexLanguageModel extends BaseLanguageModel {
       throw new RuntimeException("Requested " + tokens.size() + "gram but index has only up to " + maxNgram + "gram: " + tokens);
     }
     Objects.requireNonNull(tokens);
-    Term term = new Term("ngram", StringUtils.join(tokens, " "));
+    Term term = new Term("ngram", String.join(" ", tokens));
     return getCount(term, getLuceneSearcher(tokens.size()));
   }