+++ languagetool-core/src/main/java/org/languagetool/markup/AnnotatedText.java
@@ -18,7 +18,7 @@
  */
 package org.languagetool.markup;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 
 import java.util.List;
 import java.util.Map;