+++ languagetool-wikipedia/src/main/java/org/languagetool/dev/wikipedia/SwebleWikipediaTextFilter.java
@@ -18,7 +18,7 @@
  */
 package org.languagetool.dev.wikipedia;
 
-import org.apache.commons.lang.StringUtils;
+import org.apache.commons.lang3.StringUtils;
 import org.sweble.wikitext.engine.CompiledPage;
 import org.sweble.wikitext.engine.Compiler;
 import org.sweble.wikitext.engine.PageId;